+++ common/src/main/java/com/jannik_kuehn/common/storage/file/FileNameStorage.java
@@ -82,6 +82,18 @@ public class FileNameStorage implements NameStorage {
     }
 
     @Override
+    public void removeUser(final UUID uniqueId) throws StorageException {
+        Objects.requireNonNull(uniqueId);
+        getName(uniqueId).ifPresent(name -> {
+            try {
+                storageProvider.delete(name);
+            } catch (final StorageException e) {
+                throw new RuntimeException(e);
+            }
+        });
+    }
+
+    @Override
     public void close() throws StorageException {
         storageProvider.close();
     }